HAR-76: Refactor code.
